more on properties

+++ property/inc/class.boworkorder.inc.php
@@ -561,7 +561,7 @@
 			{
 				$entry['user_lid'] = $GLOBALS['phpgw']->accounts->id2name($GLOBALS['phpgw']->accounts->name2id($entry['user_lid']));
 				$entry['tender_delay']	 = phpgwapi_datetime::get_working_days($entry['tender_deadline'], $entry['tender_received']);
-				$entry['end_date_delay'] = round(phpgwapi_datetime::get_working_days($entry['end_date'], $entry['inspection_on_completion'] ? $entry['inspection_on_completion'] : time()));
+				$entry['end_date_delay'] = round((int)phpgwapi_datetime::get_working_days($entry['end_date'], $entry['inspection_on_completion'] ? $entry['inspection_on_completion'] : time()));
 
 				//Formatting
 				$entry['entry_date']				 = $GLOBALS['phpgw']->common->show_date($entry['entry_date'], $dateformat);
